// src/types/product.types.ts
export interface ProductRating {
  rate: number;
  count: number;
}

export interface Product {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
  rating: ProductRating;
}

export interface FilterState {
  searchQuery: string;
  selectedCategory: string;
  priceRange: [number, number];
  currentPage: number;
  pageSize: number;
  selectedProductId: number | null;
}